+++ pyviews/binding/__init__.py
@@ -1,8 +1,6 @@
 """Binding implementations"""
 
 from .binder import Binder, BindingRule, BindingContext
-from .implementations import PropertyTarget, FunctionTarget
-from .implementations import PropertyExpressionTarget, GlobalValueExpressionTarget
 from .implementations import ExpressionBinding, ObservableBinding, TwoWaysBinding
 from .implementations import get_expression_target
 from .rules import OnceRule, OnewayRule