+++ pyviews/binding/rules.py
@@ -1,7 +1,9 @@
 """Binding rules and factory"""
+from functools import partial
+
 from pyviews.binding.binder import BindingRule, BindingContext
 from pyviews.compilation import Expression, execute
-from .implementations import ExpressionBinding, PropertyTarget, InlineBinding
+from .implementations import ExpressionBinding, InlineBinding
 from ..core import Binding
 
 
@@ -22,10 +24,10 @@ class OnewayRule(BindingRule):
     def suitable(self, _: BindingContext) -> bool:
         return True
 
-    def apply(self, context: BindingContext):
+    def apply(self, context: BindingContext) -> Binding:
         expr = Expression(context.expression_body)
-        target = PropertyTarget(context.node, context.xml_attr.name, context.modifier)
-        binding = ExpressionBinding(target, expr, context.node.node_globals)
+        on_update = partial(context.modifier, context.node, context.xml_attr.name)
+        binding = ExpressionBinding(on_update, expr, context.node.node_globals)
         binding.bind()
         return binding
 
@@ -40,7 +42,7 @@ class InlineRule(BindingRule):
         (bind_body, value_body) = context.expression_body.split('}:{')
         bind_expr = Expression(bind_body)
         value_expr = Expression(value_body)
-        target = PropertyTarget(context.node, context.xml_attr.name, context.modifier)
-        binding = InlineBinding(target, bind_expr, value_expr, context.node.node_globals)
+        on_update = partial(context.modifier, context.node, context.xml_attr.name)
+        binding = InlineBinding(on_update, bind_expr, value_expr, context.node.node_globals)
         binding.bind()
         return binding